Move ByteArray from utils to utils.extensions
