+++ app/src/main/kotlin/com/absinthe/libchecker/utils/extensions/ByteArrayExtensions.kt
@@ -1,4 +1,4 @@
-package com.absinthe.libchecker.utils
+package com.absinthe.libchecker.utils.extensions
 
 fun ByteArray.toHexString(separator: CharSequence = "") = joinToString(separator) {
   it.toInt().and(0xff).toString(16).padStart(2, '0')
